TestIHSodin/requirements.txt
arrow==0.10.0
astroid==1.5.2
backports.functools-lru-cache==1.3
BeautifulSoup==3.2.1
binaryornot==0.4.3
chardet==3.0.2
click==6.7
colorama==0.3.7
configparser==3.5.0
cookiecutter==1.5.1
decorator==4.0.11
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
geocoder==1.21.2
httplib2==0.10.3
isort==4.2.5
jdcal==1.3
Jinja2==2.9.6
jinja2-time==0.2.0
lazy-object-proxy==1.2.2
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.1
pager==3.3
poyo==0.4.1
pylint==1.7.1
pymongo==3.3.0
python-dateutil==2.6.0
python-instagram==1.3.2
ratelim==0.1.6
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
twython==3.4.0
whichcraft==0.4.0
wrapt==1.10.10
numpy==1.12.1
